Counter VDU now shows.  I do not really know what the Counter VDU really showed, so I decided to add an instruction counter to the Cpu and show the insructions executed on the Counter VDU.
At this point the VDUs are pretty much complete.  From video I
have seen the register VDUs showed the waveform as square waves,
as if viewed on an oscilloscope, I have emulated this format.  For
multi line registers displays, the accumulator and multiplier
registers, I do not know exactly how they were displayed, so for
now they appear as two lines on the respective VDUs